+++ index.ts
@@ -1,4 +1,1 @@
-import sessionCleanupExtension from "./src/index.js";
-
 export { apply, inject, name } from "./src/dsh/plugin.js";
-export default sessionCleanupExtension;